backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
ConfigArgParse==1.7.1
diskcache==5.6.3
fastapi==0.135.1
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.9
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
llama_cpp_python==0.3.16
locust==2.43.3
MarkupSafe==3.0.3
msgpack==1.1.2
numpy==2.4.2
packaging==26.0
pluggy==1.6.0
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-engineio==4.13.1
python-socketio==5.16.1
pyzmq==27.1.0
requests==2.32.5
simple-websocket==1.1.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
wsproto==1.3.2
zope.event==6.1
zope.interface==8.2
